add c-styled header file to curve-sdk
